+++ src/floobits/windows/ChatForm.java
@@ -1,27 +1,17 @@
 package floobits.windows;
 
-import com.intellij.openapi.application.ApplicationManager;
-import com.intellij.openapi.ui.popup.Balloon;
-import com.intellij.openapi.ui.popup.JBPopupFactory;
 import com.intellij.ui.JBColor;
-import com.intellij.ui.awt.RelativePoint;
 import com.intellij.ui.components.JBScrollPane;
-import com.intellij.uiDesigner.core.GridConstraints;
-import com.intellij.util.ui.GridBag;
 import floobits.common.interfaces.IContext;
-import floobits.common.RunLater;
 import floobits.common.protocol.handlers.FlooHandler;
-import floobits.common.protocol.FlooUser;
-import floobits.dialogs.SetPermissionsDialog;
 import floobits.impl.ContextImpl;
 import floobits.utilities.Colors;
 import floobits.utilities.Flog;
+import floobits.windows.ChatUserForm.ClientModelItem;
 
 import javax.swing.*;
 import javax.swing.event.HyperlinkEvent;
 import javax.swing.event.HyperlinkListener;
-import javax.swing.event.PopupMenuEvent;
-import javax.swing.event.PopupMenuListener;
 import javax.swing.text.BadLocationException;
 import javax.swing.text.html.HTMLDocument;
 import javax.swing.text.html.HTMLEditorKit;
@@ -31,51 +21,11 @@ import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.text.SimpleDateFormat;
-import java.util.Date;
 import java.util.List;
+import java.util.Date;
 
 public class ChatForm {
 
-    protected class ClientChatActionListener implements ActionListener {
-
-        @Override
-        public void actionPerformed(ActionEvent e) {
-            FlooHandler flooHandler = context.getFlooHandler();
-            if (flooHandler == null) {
-                return;
-            }
-        }
-
-        protected void clientActionPerformed(FlooHandler flooHandler, ClientModelItem client) {
-
-        }
-    }
-
-    static public class ClientModelItem {
-        public String username;
-        public String client;
-        public String platform;
-        public String gravatar;
-        public int userId;
-        public Boolean following;
-
-        public ClientModelItem(String username, String gravatar, String client, String platform, Integer userId, Boolean following) {
-            this.username = username;
-            this.client = client;
-            this.platform = platform;
-            this.userId = userId;
-            this.following = following;
-            this.gravatar = gravatar;
-        }
-
-        public String toString() {
-            String formattedUsername = username;
-            if (following) {
-                formattedUsername += "*";
-            }
-            return String.format("<html><b>%s</b> <small><i>%s (%s)</html></i></small>", formattedUsername, client, platform);
-        }
-    }
 
     private JPanel chatPanel;
     private JScrollPane messagesScrollPane;
@@ -92,7 +42,6 @@ public class ChatForm {
     public ChatForm (IContext context) {
         super();
         this.context = (ContextImpl) context;
-        setupPopupMenu();
         kit = new HTMLEditorKit();
         doc = new HTMLDocument();
         messages.setEditorKit(kit);
@@ -154,13 +103,11 @@ public class ChatForm {
     }
 
     public void addUser(List<ClientModelItem> clients) {
-        ClientModelItem firstClient = clients.get(0);
+        ChatUserForm.ClientModelItem firstClient = clients.get(0);
         if (firstClient == null) {
             return;
         }
-        ChatUserForm user = new ChatUserForm();
-        user.setUsername(firstClient.username);
-        ChatUserForm userForm = new ChatUserForm();
+        ChatUserForm userForm = new ChatUserForm(context);
         userForm.setUsername(firstClient.username);
         if (firstClient.gravatar != null) {
             ContextImpl.BalloonState balloonState = context.gravatars.get(firstClient.gravatar);
@@ -168,111 +115,10 @@ public class ChatForm {
                 userForm.addGravatar(balloonState.largeGravatar, firstClient.username);
             }
         }
-        for (ClientModelItem client : clients) {
+        for (ChatUserForm.ClientModelItem client : clients) {
             userForm.addClient(client.client, client.platform);
         }
         clientsPane.add(userForm.getContainerPanel());
-    }
-
-    private void setFollowState(FlooHandler flooHandler, String msg) {
-        flooHandler.state.setFollowedUsers(flooHandler.state.followedUsers);
-        flooHandler.context.setUsers(flooHandler.state.users);
-        flooHandler.context.statusMessage(msg);
-    }
-
-    private void setupPopupMenu() {
-        final JPopupMenu popupMenu = new JPopupMenu();
-        final JMenuItem kickMenuItem = new JMenuItem("Kick");
-        kickMenuItem.addActionListener(new ClientChatActionListener() {
-            @Override
-            public void clientActionPerformed(FlooHandler flooHandler, ClientModelItem client) {
-                Flog.info("Kicking %s with user id %d.", client.username, client.userId);
-                flooHandler.editorEventHandler.kick(client.userId);
-            }
-        });
-        popupMenu.add(kickMenuItem);
-        final JMenuItem followMenuItem = new JMenuItem("Follow");
-        followMenuItem.addActionListener(new ClientChatActionListener() {
-            @Override
-            public void clientActionPerformed(FlooHandler flooHandler, ClientModelItem client) {
-                Flog.info("Following %s with user id %d.", client.username, client.userId);
-                if (flooHandler.state.followedUsers.contains(client.username)) {
-                    flooHandler.context.errorMessage(String.format("You are already following %s", client.username));
-                    return;
-                }
-                flooHandler.state.followedUsers.add(client.username);
-                setFollowState(flooHandler, String.format("You are now following %s", client.username));
-            }
-        });
-        popupMenu.add(followMenuItem);
-        final JMenuItem unFollowMenuItem = new JMenuItem("Stop following");
-        unFollowMenuItem.addActionListener(new ClientChatActionListener() {
-            @Override
-            public void clientActionPerformed(FlooHandler flooHandler, ClientModelItem client) {
-                Flog.info("Will stop following %s with user id %d.", client.username, client.userId);
-                if (!flooHandler.state.followedUsers.contains(client.username)) {
-                    flooHandler.context.errorMessage(String.format("You are not following %s", client.username));
-                    return;
-                }
-                flooHandler.state.followedUsers.remove(client.username);
-                setFollowState(flooHandler, String.format("You have stopped following %s", client.username));
-            }
-        });
-        popupMenu.add(unFollowMenuItem);
-        final JMenuItem adminMenuItem = new JMenuItem("Edit Permissions...");
-        adminMenuItem.addActionListener(new ClientChatActionListener() {
-            @Override
-            public void clientActionPerformed(FlooHandler flooHandler, ClientModelItem client) {
-                final int userId = client.userId;
-                Flog.info("Opening up permission dialog for %s", client.username);
-                FlooUser user = flooHandler.state.getUser(client.userId);
-                if (user == null) {
-                    return;
-                }
-                List<String> permissions = java.util.Arrays.asList(user.perms);
-                SetPermissionsDialog setPermissionsDialog = new SetPermissionsDialog(
-                        new RunLater<String[]>() {
-                            @Override
-                            public void run(String[] permissions) {
-                                Flog.info("Submitting permission changes.");
-                                FlooHandler flooHandler = context.getFlooHandler();
-                                if (flooHandler == null) {
-                                    return;
-                                }
-                                flooHandler.editorEventHandler.changePerms(userId, permissions);
-                            }
-                        },
-                        permissions.contains("get_buf"),
-                        permissions.contains("request_perms"),
-                        permissions.contains("patch"),
-                        permissions.contains("kick")
-                );
-                setPermissionsDialog.setUsername(client.username);
-                setPermissionsDialog.createCenterPanel();
-                setPermissionsDialog.show();
-            }
-        });
-        popupMenu.add(adminMenuItem);
-        popupMenu.addPopupMenuListener(new PopupMenuListener() {
-            @Override
-            public void popupMenuWillBecomeVisible(PopupMenuEvent e) {
-                FlooHandler floohandler = context.getFlooHandler();
-                if (floohandler == null) {
-                    return;
-                }
-                kickMenuItem.setEnabled(floohandler.state.can("kick"));
-                adminMenuItem.setEnabled(floohandler.state.can("kick"));
-
-            }
-
-            @Override
-            public void popupMenuWillBecomeInvisible(PopupMenuEvent e) {
-            }
-
-            @Override
-            public void popupMenuCanceled(PopupMenuEvent e) {
-            }
-        });
     }
 
     private void sendChatContents() {
